# swagger_documentation.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Drogaleste API",
    "description": "Documentação da API da Drogaleste"
  },
  "host": "apidrogaleste.ddns.net:7150",
  "basePath": "api/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "description": "JWT token",
      "type": "Bearer",
      "in": "body",
      "name": "Authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/user/new": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "inscricaofederal",
            "in": "header",
            "type": "string"
          },
          {
            "name": "email",
            "in": "header",
            "type": "string"
          },
          {
            "name": "nome",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "406": {
            "description": "Not Acceptable"
          }
        }
      }
    },
    "/api/user/authorize/{clientId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/drogaleste/client/purchasehistoric/{cod}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cod",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/drogaleste/client/{cod}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "cod",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}